restrutured backend: move db config into databaseHelper.py, added realtimePrice.py
