+++ MySQL/MySQL_Data
@@ -1,19 +1,23 @@
 -- Insert data into the 'stocks' table
-INSERT INTO stocks (symbol, stock_name, price) VALUES
-('AAPL', 'Apple Inc.', 150.00),
-('GOOGL', 'Alphabet Inc.', 2800.00),
-('AMZN', 'Amazon.com Inc.', 3400.00);
+INSERT INTO stocks (symbol, stock_name) VALUES
+('AAPL', 'Apple Inc.'),
+('GOOGL', 'Alphabet Inc.'),
+('AMZN', 'Amazon.com Inc.'),
+('MSFT', 'Microsoft Corporation'),
+('FB', 'Facebook Inc.'),
+('TSLA', 'Tesla Inc.');
+
 
 -- Insert data into the 'portfolio' table
-INSERT INTO portfolio (user_name, email, balance, total_value) VALUES
-('Mike Liu', 'random.rd@random.com', 10000.00, 15500.00);
+-- INSERT INTO portfolio (user_name, email, balance, total_value) VALUES
+-- ('Mike Liu', 'random.rd@random.com', 10000.00, 0.00);
 
 -- Insert data into the 'portfolio_stocks' table
-INSERT INTO portfolio_stocks (stock_name, portfolio_id, stock_id, quantity) VALUES
+-- INSERT INTO portfolio_stocks (stock_name, symbol, portfolio_id, stock_id, quantity) VALUES
 ('Apple Inc.', 1, 1, 10),
 ('Alphabet Inc.', 1, 2, 5);
 
 -- Insert data into the 'transactions' table
-INSERT INTO transactions (portfolio_id, stock_id, transaction_type, quantity) VALUES
-(1, 1, 'buy', 10),
-(1, 2, 'buy', 5);
+-- INSERT INTO transactions (portfolio_id, stock_id, symbol, transaction_type, quantity) VALUES
+(1, 1, 'AAPL', 'buy', 10),
+(1, 2, 'GOOGL', 'buy', 5);
