+++ MySQL/MySQL_Tables
@@ -1,8 +1,7 @@
 create table stocks (
 	id int auto_increment primary key,
     symbol varchar(10) unique not null,
-    stock_name varchar(100) not null,
-    price decimal(10, 2) not null
+    stock_name varchar(100) not null
 );
 
 create table portfolio (
@@ -16,18 +15,20 @@ create table portfolio (
 create table portfolio_stocks (
 	id int auto_increment primary key,
 	stock_name varchar(100) not null,
+	symbol varchar(10) unique not null,
     portfolio_id int,
     stock_id int,
     quantity int,
     foreign key (stock_id) references stocks(id),
     foreign key (portfolio_id) references portfolio(id),
-    unique (portfolio_id, stock_id)
+	unique (portfolio_id, stock_id)
 );
 
 create table transactions (
 	id int auto_increment primary key,
     portfolio_id int,
     stock_id int,
+	symbol varchar(10) unique not null,
     transaction_type enum ('buy', 'sell'),
     quantity int,
     transaction_date timestamp default current_timestamp,
